+++ src/assets/mhrise/mhrise-charm-decorations.js
