+++ src/assets/mhrise/mhrise-charm-manager.js
