+++ src/components/nav/NavOptions.svelte
@@ -1,11 +1,11 @@
 <script context="module">
-  import CharmScanner          from './CharmScanner.svelte'
-  import CharmRealtimeScanner  from './CharmRealtimeScanner.svelte'
-  import CharmList             from './CharmList.svelte'
-  import CharmSearcher         from './CharmSearcher.svelte'
-  import CharmRecycleSupporter from './CharmRecycleSupporter.svelte'
-  import CharmIxporter         from './CharmIxporter.svelte'
-  import Description           from './Description.svelte'
+  import CharmScanner          from '../../pages/CharmScanner.svelte'
+  import CharmRealtimeScanner  from '../../pages/CharmRealtimeScanner.svelte'
+  import CharmList             from '../../pages/CharmList.svelte'
+  import CharmSearcher         from '../../pages/CharmSearcher.svelte'
+  import CharmRecycleSupporter from '../../pages/CharmRecycleSupporter.svelte'
+  import CharmIxporter         from '../../pages/CharmIxporter.svelte'
+  import Description           from '../../pages/Description.svelte'
 
   export const navOptions = [
 		{ tabTitle: '護石スキャン',
@@ -30,6 +30,12 @@
       component: CharmSearcher,
       iconData: [
         "M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z",
+      ],
+    },
+    { tabTitle: '輪廻サポート',
+      component: CharmRecycleSupporter,
+      iconData: [
+        "M9.302 1.256a1.5 1.5 0 0 0-2.604 0l-1.704 2.98a.5.5 0 0 0 .869.497l1.703-2.981a.5.5 0 0 1 .868 0l2.54 4.444-1.256-.337a.5.5 0 1 0-.26.966l2.415.647a.5.5 0 0 0 .613-.353l.647-2.415a.5.5 0 1 0-.966-.259l-.333 1.242-2.532-4.431zM2.973 7.773l-1.255.337a.5.5 0 1 1-.26-.966l2.416-.647a.5.5 0 0 1 .612.353l.647 2.415a.5.5 0 0 1-.966.259l-.333-1.242-2.545 4.454a.5.5 0 0 0 .434.748H5a.5.5 0 0 1 0 1H1.723A1.5 1.5 0 0 1 .421 12.24l2.552-4.467zm10.89 1.463a.5.5 0 1 0-.868.496l1.716 3.004a.5.5 0 0 1-.434.748h-5.57l.647-.646a.5.5 0 1 0-.708-.707l-1.5 1.5a.498.498 0 0 0 0 .707l1.5 1.5a.5.5 0 1 0 .708-.707l-.647-.647h5.57a1.5 1.5 0 0 0 1.302-2.244l-1.716-3.004z",
       ],
     },
     { tabTitle: '護石エクスポート',
