+++ src/components/parts/CameraReader.svelte
