+++ src/components/parts/SvelteTable.svelte
