+++ src/components/parts/VideoReader.svelte
